Renames remaining references to game.manager.nomic to nomic package schema.
