+++ app/src/androidTest/java/mobile/nomic/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package mobile.game.manager.nomic
+package mobile.nomic
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("mobile.game.manager.nomic", appContext.packageName)
+        assertEquals("mobile.nomic", appContext.packageName)
     }
 }
